refactor(pool-page): extract _tabs/liquidity-tab
Move LiquidityTab (~200 lines) out of page.tsx into
_tabs/liquidity-tab.tsx.

Behavior-preserving — verified via:
- pnpm test (1366/1366 green)
- pnpm typecheck (clean)
- chrome-devtools snapshot of /pool/<id>?tab=liquidity vs baseline:
  zero structural changes, console clean

page.tsx: 2008 → 1808 lines (-200).
